+++ flow-typed/express.js
@@ -1,9 +1,0 @@
-// @flow
-
-declare class express$Request extends http$IncomingMessage {}
-
-declare type express$Middleware = (
-  req: express$Request,
-  res: any,
-  next: any
-) => mixed