+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './types.js'
 export * from './middleware/routing.js'
 export * from './middleware/with-cors.js'
+export * from './adapters/express.js'
 export * from './adapters/fetch.js'